Deprecate the `diesel_` prefixed macros
In Rust 2018 you can write `diesel::some_macro!`, removing the need for
the `diesel_` prefix on these generically named macros. Note that this
commit by itself does not make these macros usable in that form, we
either need to replace all `#[macro_export]` with
`#[macro_export(local_inner_macros)]` or prefix all macro invocations
with `$crate::` for that to be possible. The latter is planned prior to
the 2.0 release.

I want to do a similar deprecation for `#[derive(DieselNumericOps)` and
`#[diesel(...)]`, but those require a bit more investigation/design.