remove confusing panic messages from examples
The `publish_post` examples contained a confusing `panic!()` that
suggested the update statement would return an error, when a post to be
updated could not be found. In reality, however, that case results in an
`Ok(0)` value.

To avoid confusion, this commit replaces the panicking
`unwrap_or_else()` calls with a simpler `unwrap()`.