Handle on conflict clauses for insert from select statements on sqlite
Sqlite is not able to handle queries with the following scheme:
`INSERT INTO t1 SELECT * FROM t2 ON CONFLICT(x) DO UPDATE SET
y=excluded.y;` because of a parsing ambiguity. A workaround here is
to add a dummy `WHERE 1=1` clause to the select statement.
This commit implements the following behaviour:
* Reject insert from select statements without where clause and with
on conflict clause on sqlite if the select clause is a
`SelectStatement` from diesel. (Read: If it's not boxed and we can
tell that at compile time)
* For boxed statements dynamically add a `WHERE 1=1` if no where
clause exists (Only on sqlite and for Insert from select statements
with on conflict clause)

See their documentation for details and workarounds: https://sqlite.org/lang_UPSERT.html