Remove unused metadata field from `PgValue`
This was originally included in #2134 with the following comment:

> Additionally some type information are added to the postgresql value
> type. Those information may allow to do dynamic type conversations in
> future diesel versions.

While this may be useful in the future, we should wait until we have a
concrete use case for doing so. At the moment this means we have to
thread a `PgConnection` through various types that otherwise wouldn't
need to carry it, and are stabilizing an API that we aren't even sure
if/how we want to eventually use. Given that every use of `PgConnection`
complicates providing an async implementation, we should be more
conservative about speculative APIs like this.