Use unnamed free constants to wrap derive impls
As of Rust 1.37, we no longer need to come up with unique names to use
for everything, we can just use an unnamed constant. We originally
switched away from wrapping in a constant to work around a bug in Rust
that appears to have since been fixed.

I suspect there are some old workarounds floating in the codebase that
we can remove with this change, but none immediately stood out to me, so
they can be removed the next time they get touched (if at all)