Fix travis build
The build on master is currently broken. This is because the lockfile
created by Cargo in 1.41 is incompatible with earlier versions of Cargo.
Because the `test_all` files in the example directory aren't adding
`+$TRAVIS_RUST_VERSION`, our examples get run with the version specified
in `rust-toolchain` instead of the intended version for that piece of
the matrix.

Since everything in the main workspace uses a single lockfile, this
means that beta started breaking. We run the tests for the main crate
and derives first, which (correctly) runs on 1.41 and generates a
lockfile. When we then try to run the examples, they (incorrectly) use
the version in `rust-toolchain`, which cannot read the lockfile
generated by 1.41.

To fix this I just nuke the file before the build starts.