Generalize creating a `BoxedWhereClause`
This commit generalizes the creating of a boxed where clause from an
empty where clause in such a way that this works for any lifetime
instead only for static lifetimes. This is required to write
`dsl::IntoBoxed<some::table>` for general queries that may contain a
restriction on their lifetime. (For example because of they contain a
`.eq_any(some_borrowed_array)` expression)