Add a way to lookup return bind types in the mysql backend
For supporting dynamic select statements we need a way to say: We
don't know the type of the select clause. This change adds a
possibility to say, We don't know which mysql field type is returned
here, by setting the type metadata to none. There is only one
reasonable case to do this (to implement a marker type for dynamic
select clauses), so we've updated the custom derive to unconditionally
insert a specific value there.