Fix transactions not closing properly when failing on commit
Connection could be made available again with a still-open transaction due to transaction commit fail (typically `SerializationFailure` or `ReadOnlyTransaction`).

This eventually triggered "AlreadyInTransaction: Cannot perform this operation while a transaction is open" errors in unrelated places.

This commit fixes this issue by catching `SerializationFailure` or `ReadOnlyTransaction` on commit and attempting a rollback.