Fix #2373
This change fixes a out of buffer write in newer versions of
libmysqlclient. That write was caused by a change in libmysqlclient
around version 8.0.19 that added another field to `MYSQL_TIME` and by
diesel using a statically generated binding (mysqlclient-sys). As
consequence diesel would allocate a to small buffer for the struct
while libmysql writes more bytes that expected, which results in a out
of bound write. Interestingly this does not happen with a recent
libmysqlclient version based on the sources in the mariadb source
code.

As a fix I move an updated version of `MYSQL_TIME` to diesel itself, as I do not have any
control over `mysqlclient-sys`. This change is compatible with older
libmysqlclient versions, because for them we just allocate a larger
buffer than they actually use, which is fine.