Make `r2d2::PooledConnection` more generic
The `Connection` impl for `PooledConnection` was restricted to apply
only to `Connection` types that use `AnsiTransactionManager`, which is
not necessary for the general case. At least `diesel-oci` is using a
different Transaction manager impl, so just make this impl more
generic by introducing a transparent wrapper type that implements
`TransactionManager` for all cases.