Remove the order clause from diesels medium complex query benches
The other benchmark do also not include a order by clause, so this
queries should do the same. I've opted for removing the order clause, as
we are interested in benchmarking the framework not the databases. I
assume that an order clause results in more time spend inside the database.