Fix type confusion in custom type lookup
The old code did not incorporate the order of schemas in your search
path. As consequence it could pick the wrong type if types with the same
name exists in more than one schema. This commit fixes that issue by
performing the name resolution on database side by casting the type name
to the type itself.