Rewrite bind serialization layer
This commit removes a old hack from our bind serialization layer that
requires multiple value copies for the sqlite backend going first to a
binary format and later on deserialize the data again to call the
corresponding bind function. After this commit it is possible to just
store the corresponding values and call the corresponding bind functions
directly onto those values. This removes a therotical overhead and
should improve the performance on bind heavy workloads (like inserts).
To do this we basically need to teach the compiler about a invariant
that is already true: Bind values need to live at least as long as the
corresponding query. To describe that in the type system we need to
introduce a few lifetime bounds for `QueryFragment::walk_ast` and
`ToSql::to_sql`. In exchange we can remove the write bound from
`ToSql::to_sql`. In my opion that does not make the it harder to
implement the corresponding traits as users don't need to interact with
those lifetimes at all and they describe something that is already true
for almost all cases. Additionally now that rust-analyzer can generate
function bodies for trait implementations on the fly users don't need to
type those bounds, as they are generated with the body stub.
Most of the changes are releated to just introduce the lifetime bounds
in all nessesary impls and fix minor incompatibilities there.